+++ README.md
@@ -26,10 +26,11 @@ information on the current packages:
 
 ## npm packages
 
-| Package                                                                          |                                                             Security Status                                                              |
-| -------------------------------------------------------------------------------- | :--------------------------------------------------------------------------------------------------------------------------------------: |
-| [mdx-embed](https://www.npmjs.com/package/mdx-embed)                             |               [![Known Vulnerabilities](https://snyk.io/test/npm/mdx-embed/badge.svg)](https://snyk.io/test/npm/mdx-embed)               |
-| [gatsby-plugin-mdx-embed](https://www.npmjs.com/package/gatsby-plugin-mdx-embed) | [![Known Vulnerabilities](https://snyk.io/test/npm/gatsby-plugin-mdx-embed/badge.svg)](https://snyk.io/test/npm/gatsby-plugin-mdx-embed) |
+| Package                                                                              |                                                               Security Status                                                                |
+| ------------------------------------------------------------------------------------ | :------------------------------------------------------------------------------------------------------------------------------------------: |
+| [mdx-embed](https://www.npmjs.com/package/mdx-embed)                                 |                 [![Known Vulnerabilities](https://snyk.io/test/npm/mdx-embed/badge.svg)](https://snyk.io/test/npm/mdx-embed)                 |
+| [gatsby-plugin-mdx-embed](https://www.npmjs.com/package/gatsby-plugin-mdx-embed)     |   [![Known Vulnerabilities](https://snyk.io/test/npm/gatsby-plugin-mdx-embed/badge.svg)](https://snyk.io/test/npm/gatsby-plugin-mdx-embed)   |
+| [storybook-addon-mdx-embed](https://www.npmjs.com/package/storybook-addon-mdx-embed) | [![Known Vulnerabilities](https://snyk.io/test/npm/storybook-addon-mdx-embed/badge.svg)](https://snyk.io/test/npm/storybook-addon-mdx-embed) |
 
 <br />
