+++ docs/pages/1.introduction.stories.mdx
@@ -1,40 +1,30 @@
 import { Meta } from '@storybook/addon-docs/blocks';
 import { linkTo } from '@storybook/addon-links';
 
+import { Flex, Box, Grid, Link } from 'theme-ui';
+import { GitHubLink, Quote } from '../src/components';
+
 <Meta title="Introduction" />
 
-<Flex sx={{ alignItems: 'center' }}>
-  <Heading as="h1" sx={{ flexGrow: 1 }}>
-    MDX Embed
-  </Heading>
-  <Link
-    href="https://github.com/PaulieScanlon/mdx-embed"
-    target="_blank"
-    sx={{
-      color: 'text',
-    }}
-  >
-    <IconButton tabIndex={-1}>
-      <svg xmlns="http://www.w3.org/2000/svg" role="img" viewBox="0 0 24 24" width="24" height="24" fill="currentColor">
-        <title>GitHub icon</title>
-        <g>
-          <path d="M12 .297c-6.63 0-12 5.373-12 12 0 5.303 3.438 9.8 8.205 11.385.6.113.82-.258.82-.577 0-.285-.01-1.04-.015-2.04-3.338.724-4.042-1.61-4.042-1.61C4.422 18.07 3.633 17.7 3.633 17.7c-1.087-.744.084-.729.084-.729 1.205.084 1.838 1.236 1.838 1.236 1.07 1.835 2.809 1.305 3.495.998.108-.776.417-1.305.76-1.605-2.665-.3-5.466-1.332-5.466-5.93 0-1.31.465-2.38 1.235-3.22-.135-.303-.54-1.523.105-3.176 0 0 1.005-.322 3.3 1.23.96-.267 1.98-.399 3-.405 1.02.006 2.04.138 3 .405 2.28-1.552 3.285-1.23 3.285-1.23.645 1.653.24 2.873.12 3.176.765.84 1.23 1.91 1.23 3.22 0 4.61-2.805 5.625-5.475 5.92.42.36.81 1.096.81 2.22 0 1.606-.015 2.896-.015 3.286 0 .315.21.69.825.57C20.565 22.092 24 17.592 24 12.297c0-6.627-5.373-12-12-12" />
-        </g>
-      </svg>
-    </IconButton>
-  </Link>
+<Flex sx={{ alignItems: 'center', justifyContent: 'space-between' }}>
+
+
+# MDX Embed
+
+<GitHubLink />
+
 </Flex>
 
-MDX Embed allows you to easily _embed_ popular 3rd party media content such as YouTube videos, Tweets, Instagram posts, Egghead lessons, Spotify, TikTok and many more straight into your `.mdx` - **no import required!**
 
-<Divider />
+MDX Embed allows you to easily _embed_ popular 3rd party media content such as YouTube videos, Tweets, Instagram posts, Egghead lessons, Spotify, TikTok and many more straight into your `.mdx` - **no import required!**
 
 There are a couple of ways to use MDX Embed so you may wish to explore each of the packages:
 
 - [mdx-embed](/?path=/story/mdx-embed--page)
 - [gatsby-plugin-mdx-embed](/?path=/story/gatsby-plugin--page)
+- [storybook-addon-mdx-embed](/?path=/story/storybook-addon--page)
 
-<Divider sx={{ my: 4 }} />
+<Box sx={{ height: '24px' }} />
 
 ## How it works
 
@@ -51,15 +41,13 @@ Here's a pen, and some other blog post text
 
 ## _Which renders as 👇_
 
-<Divider sx={{ my: 2 }} />
-
 #### My cool pen
 
 Here's a pen, and some other blog post text
 
 <CodePen codePenId="PNaGbb" />
 
-<Box sx={{ height: '84px' }} />
+<Box sx={{ height: '24px' }} />
 
 ## What others say
 
@@ -71,7 +59,7 @@ Here's a pen, and some other blog post text
   />
 </Grid>
 
-<Divider sx={{ my: 4 }} />
+<Box sx={{ height: '24px' }} />
 
 ## The team
 
@@ -79,7 +67,7 @@ Here's a pen, and some other blog post text
 - Scott Spence | [@spences10](https://twitter.com/spences10)
 - Rich Haines | [@studio_hungry](https://twitter.com/studio_hungry)
 
-<Divider sx={{ my: 4 }} />
+<Box sx={{ height: '24px' }} />
 
 ### Legacy Contributors
