+++ docs/pages/2.mdx-embed.stories.mdx
@@ -2,7 +2,7 @@ import { Meta } from '@storybook/addon-docs/blocks';
 
 <Meta title="mdx-embed" />
 
-## MDX Embed Install
+# MDX Embed Install
 
 ```sh
 npm install mdx-embed --save
