+++ docs/pages/3.gatsby-plugin.stories.mdx
@@ -2,7 +2,7 @@ import { Meta } from '@storybook/addon-docs/blocks';
 
 <Meta title="gatsby-plugin" />
 
-## Gatsby Plugin Install
+# Gatsby Plugin Install
 
 ```sh
 npm install mdx-embed gatsby-plugin-mdx-embed --save
@@ -24,6 +24,7 @@ module.exports = {
 ### Alternatively
 
 The plugin works by using Gatsby's `wrapRootElement` and returning the `MDXEmbedProvider`.
+
 If you encounter any issues with the plugin you could try manually including the `MDXEmbedProvider` in your project.
 Ensure you have removed the plugin from your dependencies and plugins array in `gatsby-config.js` then add the following to `gatsby-browser.js` and / or `gatsby-ssr.js`
