+++ docs/pages/components/egghead/egghead-lesson.stories.mdx
