+++ docs/src/components/index.js
@@ -1,0 +1,2 @@
+export { GitHubLink } from './github-link';
+export { Quote } from './quote';
