+++ packages/storybook-addon-mdx-embed/babel.config.json
@@ -1,0 +1,11 @@
+{
+  "presets": ["@babel/env", "@babel/preset-react"],
+  "plugins": [
+    [
+      "@babel/plugin-transform-modules-commonjs",
+      {
+        "strictMode": false
+      }
+    ]
+  ]
+}
