+++ packages/storybook-addon-mdx-embed/src/index.js
@@ -1,0 +1,1 @@
+// boop
