+++ packages/storybook-addon-mdx-embed/src/preset/index.js
@@ -1,0 +1,3 @@
+export function config(entry = []) {
+  return [...entry, require.resolve('./default-parameters')];
+}
